Clean up

+++ sync-terminal.sh
@@ -1,4 +1,8 @@
 # Copy everything into the filesystem
 yes | cp -R files/. ~/
 yes | cp -R color-scheme/solarized/vim-colors-solarized/colors/. ~/.vim/colors/
-yes | cp .env ~/.env.local
+
+# Copy local environment variables
+test -e .env && yes | cp .env ~/.env.local
+
+echo "Files copied successfully"
